refactor(🚚): move resources #21
Adjusted root namespace for the DomainCore.Mediator project. The Mediator library in the demo project is throwing errors which I suspect might be because of a namespace conflict. I suspect that the source generators in the Mediator library only look for namespaces ending with 'Mediator' rather than qualifying the entire namespace.

Signed-off-by: Justinian Erdmier <48148443+JustinianErdmier@users.noreply.github.com>